#360 New stocklevel trait, update to appserviceprovider and update to bom model
